gcp_minio_dedup_cf/requirements.txt
google-cloud-storage
pytest
pytest-mock
functions-framework
